rabbitmq/requirements.txt
pika
kombu
python-dotenv
